Move isTalkbackActive to a11y package.
