+++ libraries/ui-utils/src/main/kotlin/io/element/android/libraries/ui/utils/a11y/IsTalkbackEnabled.kt
@@ -6,7 +6,7 @@
  * Please see LICENSE files in the repository root for full details.
  */
 
-package io.element.android.libraries.ui.utils.time
+package io.element.android.libraries.ui.utils.a11y
 
 import android.view.accessibility.AccessibilityManager
 import androidx.compose.runtime.Composable
